Fail closed on stale experiment assignment and isolate overlay values.
Skip corrupt RUNNING snapshots instead of 500ing every evaluate/bootstrap
request. Keep frozen experiment values on the overlay rule so published
variants still apply to holdout, fallback, and non-experiment traffic.
When server bootstrap is stale, ignore overlay rules rather than assigning
experiment traffic that cannot mint an exposure token.
