+++ packages/shared/src/feature-flags/evaluator.ts
@@ -277,6 +277,7 @@ export function evaluateFeatureFlag(
   config: FeatureFlagsConfig,
   context: FeatureFlagEvaluationContext,
   seenFlagIds: ReadonlySet<string> = new Set(),
+  ignoreExperimentAssignments = false,
 ): FeatureFlagEvaluationResult {
   const flag = config.flags?.[flagId];
   if (flag === undefined) return { flagId, flagKey: flagId, reason: "missing" };
@@ -289,7 +290,7 @@ export function evaluateFeatureFlag(
   nextSeen.add(flagId);
   for (const prerequisite of Object.values(flag.prerequisites ?? {})) {
     if (prerequisite === undefined || prerequisite.flagId === undefined) return defaultResult(flagId, flag, "prerequisite_unmet");
-    const evaluated = evaluateFeatureFlag(prerequisite.flagId, config, context, nextSeen);
+    const evaluated = evaluateFeatureFlag(prerequisite.flagId, config, context, nextSeen, ignoreExperimentAssignments);
     if (evaluated.reason === "dependency_cycle") return defaultResult(flagId, flag, "dependency_cycle");
     if (evaluated.variantKey === undefined || prerequisite.variantKeys?.[evaluated.variantKey] !== true) {
       return defaultResult(flagId, flag, "prerequisite_unmet");
@@ -321,6 +322,11 @@ export function evaluateFeatureFlag(
     .sort(([leftId, left], [rightId, right]) => (right.priority ?? 0) - (left.priority ?? 0) || lexicalCompare(leftId, rightId));
   for (const [ruleId, rule] of rules) {
     if (rule.enabled === false) continue;
+    // Stale server bootstrap still contains overlay rules, but those
+    // assignments cannot mint exposure tokens. Skip them so unauditable
+    // experiment traffic is not served from a snapshot that may also be
+    // paused or completed on the backend.
+    if (ignoreExperimentAssignments && rule.experimentRunId !== undefined) continue;
     if (selectedExperimentId !== undefined && rule.experimentId !== undefined && rule.experimentId !== selectedExperimentId) continue;
     const conditions = Object.values(rule.conditions ?? {}).filter((condition) => condition !== undefined);
     if (!conditions.every((condition) => conditionMatches(condition, config, context, new Set()))) continue;
@@ -338,11 +344,12 @@ export function evaluateFeatureFlag(
       ? rule.variantKey
       : chooseFeatureFlagVariant(subjectId, `variant.${flagId}.${ruleSalt}`, weightedVariants);
     if (variantKey === undefined || flag.variants?.[variantKey] === undefined) continue;
+    const override = rule.variantValues?.[variantKey];
     return {
       flagId,
       flagKey: flag.key ?? flagId,
       variantKey,
-      value: variantValue(flag, variantKey),
+      value: override !== undefined ? override : variantValue(flag, variantKey),
       reason: "matched_rule",
       ruleId,
       ...(rule.experimentId === undefined ? {} : { experimentId: rule.experimentId }),
@@ -511,4 +518,40 @@ import.meta.vitest?.test("holdout fails closed when mixed assignment units make 
     },
   };
   expect(evaluateFeatureFlag("mixed", config, { distinctId: "user-a", userId: "user-a", teamId: "team-a" }).reason).toBe("holdout");
+});
+
+import.meta.vitest?.test("rule-local variantValues apply only to the matching experiment rule", ({ expect }) => {
+  const config: FeatureFlagsConfig = {
+    flags: {
+      copy: {
+        key: "copy",
+        enabled: true,
+        fallbackVariantKey: "control",
+        allocationSalt: "copy",
+        variants: { control: { value: "published-control" }, treatment: { value: "published-treatment" } },
+        rules: {
+          experiment: {
+            priority: 2_000_000,
+            variantWeights: { control: 5_000, treatment: 5_000 },
+            variantValues: { control: "frozen-control", treatment: "frozen-treatment" },
+            experimentId: "copyExperiment",
+            experimentRunId: "run-1",
+            stickyBy: "userId",
+          },
+          everyone: { priority: 1, variantKey: "treatment" },
+        },
+      },
+    },
+    experiments: {
+      copyExperiment: { flagId: "copy", assignmentUnit: "user" },
+    },
+  };
+  const inExperiment = evaluateFeatureFlag("copy", config, { distinctId: "user-a", userId: "user-a" });
+  expect(inExperiment.reason).toBe("matched_rule");
+  expect(inExperiment.ruleId).toBe("experiment");
+  expect(["frozen-control", "frozen-treatment"]).toContain(inExperiment.value);
+
+  const ignored = evaluateFeatureFlag("copy", config, { distinctId: "user-a", userId: "user-a" }, new Set(), true);
+  expect(ignored).toMatchObject({ reason: "matched_rule", ruleId: "everyone", value: "published-treatment" });
+  expect(ignored.experimentRunId).toBeUndefined();
 });
