+++ sdks/spec/src/apps/server-app.spec.md
@@ -26,7 +26,7 @@ Fetch `GET /api/v1/feature-flags/bootstrap` with the secret server key. The resp
 
 On a transient network, 408, 429, or 5xx failure, a previously validated snapshot may be used for at most five minutes and returned details set `isStale=true`. Never serve stale definitions for authorization/4xx failures. If no eligible snapshot exists during a transient failure, report the error through the SDK error reporter and return every caller-provided fallback with reason `error`. Other bootstrap or schema failures fail loudly.
 
-Use the shared deterministic evaluator from the feature-flags core package. Derive `distinctId`, `userId`, and trusted user attributes from `getUser({ or: "anonymous" })`. The trusted attributes are the canonical ID, verified primary email (or null), primary-email verification state, and sign-up timestamp; callers cannot override them through feature-flag context. Pass bounded non-security context and the selected team ID separately. When a team is selected, also pass `team: { id }` so `team.id` targeting matches the client evaluate endpoint. Attach the bootstrap config version to every result. Ordinary flags are fully local. An active experiment assignment may use the authenticated evaluation endpoint only to mint a signed exposure credential; the assignment returned to application code must remain the locally evaluated assignment, and a token-mint failure must not invalidate an otherwise eligible stale bootstrap result.
+Use the shared deterministic evaluator from the feature-flags core package. Derive `distinctId`, `userId`, and trusted user attributes from `getUser({ or: "anonymous" })`. The trusted attributes are the canonical ID, verified primary email (or null), primary-email verification state, and sign-up timestamp; callers cannot override them through feature-flag context. Pass bounded non-security context and the selected team ID separately. When a team is selected, also pass `team: { id }` so `team.id` targeting matches the client evaluate endpoint. Attach the bootstrap config version to every result. Ordinary flags are fully local. When bootstrap is stale, ignore live experiment overlay rules and serve ordinary targeting or the published fallback so unauditable experiment traffic is not assigned. When bootstrap is fresh, an active experiment assignment may use the authenticated evaluation endpoint only to mint a signed exposure credential; the assignment returned to application code must remain the locally evaluated assignment. A token-mint failure must not change that assignment.
 
 Feature flags are not authorization controls even when evaluated by the server SDK.
